PracticeModel, UtteanceModel 파일 따로 분리
